move search settings api to its own route
